+++ roll/testing.py
@@ -1,10 +1,13 @@
+import http
 import json
 import mimetypes
+import pytest
+import socket
+
+from functools import partial
 from io import BytesIO
 from urllib.parse import urlencode
 from uuid import uuid4
-
-import pytest
 
 
 def encode_multipart(data, charset='utf-8'):
@@ -62,12 +65,16 @@ class Transport:
 
     def __init__(self):
         self.data = b''
+        self._closing = False
+
+    def is_closing(self):
+        return self._closing
 
     def write(self, data):
         self.data += data
 
     def close(self):
-        ...
+        self._closing = True
 
 
 class Client:
@@ -129,16 +136,17 @@ class Client:
         body, headers = self.encode_body(body, headers)
         if isinstance(body, str):
             body = body.encode()
+        if body and 'Content-Length' not in headers:
+            headers['Content-Length'] = len(body)
         self.protocol = self.app.factory()
         self.protocol.connection_made(Transport())
-        self.protocol.on_message_begin()
-        self.protocol.on_url(path.encode())
-        self.protocol.request.body = body
-        self.protocol.request.method = method
-        for key, value in headers.items():
-            self.protocol.on_header(key.encode(), value.encode())
-        await self.app(self.protocol.request, self.protocol.response)
-        self.protocol.write()
+        headers = '\r\n'.join('{}: {}'.format(*h) for h in headers.items())
+        data = b'%b %b HTTP/1.1\r\n%b\r\n\r\n%b' % (
+            method.encode(), path.encode(), headers.encode(), body or b'')
+
+        self.protocol.data_received(data)
+        if self.protocol.task:
+            await self.protocol.task
         return self.protocol.response
 
     async def get(self, path, **kwargs):
@@ -178,3 +186,57 @@ def client(app, event_loop):
     app.loop.run_until_complete(app.startup())
     yield Client(app)
     app.loop.run_until_complete(app.shutdown())
+
+
+def unused_port():
+    """Return a port that is unused on the current host."""
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        s.bind(('127.0.0.1', 0))
+        return s.getsockname()[1]
+
+
+class LiveClient:
+
+    def __init__(self, app):
+        self.app = app
+        self.url = None
+        self.wsl = None
+
+    def start(self):
+        self.port = unused_port()
+        self.app.loop.run_until_complete(self.app.startup())
+        self.server = self.app.loop.run_until_complete(
+            self.app.loop.create_server(
+                self.app.factory, '127.0.0.1', self.port))
+        self.url = 'http://127.0.0.1:{port}'.format(port=self.port)
+        self.wsl = 'ws://127.0.0.1:{port}'.format(port=self.port)
+
+    def stop(self):
+        self.server.close()
+        self.port = self.url = self.wsl = None
+        self.app.loop.run_until_complete(self.server.wait_closed())
+        self.app.loop.run_until_complete(self.app.shutdown())
+
+    def execute_query(self, method, uri, headers):
+        self.conn.request(method, uri, headers=headers)
+        response = self.conn.getresponse()
+        return response
+
+    async def query(self, method, uri, headers: dict=None):
+        if headers is None:
+            headers = {}
+
+        self.conn = http.client.HTTPConnection('127.0.0.1', self.port)
+        requester = partial(self.execute_query, method.upper(), uri, headers)
+        response = await self.app.loop.run_in_executor(None, requester)
+        self.conn.close()
+        return response
+
+
+@pytest.fixture()
+def liveclient(app, event_loop):
+    app.loop = event_loop
+    client = LiveClient(app)
+    client.start()
+    yield client
+    client.stop()