+++ tests/test_response.py
@@ -24,7 +24,7 @@ async def test_can_set_status_from_httpstatus(client, app):
 
     resp = await client.get('/test')
     assert resp.status == HTTPStatus.ACCEPTED
-    assert client.protocol.writer.data == \
+    assert client.protocol.transport.data == \
         b'HTTP/1.1 202 Accepted\r\nContent-Length: 0\r\n\r\n'
 
 
@@ -36,7 +36,7 @@ async def test_write(client, app):
         resp.body = 'body'
 
     await client.get('/test')
-    assert client.protocol.writer.data == \
+    assert client.protocol.transport.data == \
         b'HTTP/1.1 200 OK\r\nContent-Length: 4\r\n\r\nbody'
 
 
@@ -47,7 +47,7 @@ async def test_write_get_204_no_content_type(client, app):
         resp.status = HTTPStatus.NO_CONTENT
 
     await client.get('/test')
-    assert client.protocol.writer.data == b'HTTP/1.1 204 No Content\r\n\r\n'
+    assert client.protocol.transport.data == b'HTTP/1.1 204 No Content\r\n\r\n'
 
 
 async def test_write_get_304_no_content_type(client, app):
@@ -57,7 +57,7 @@ async def test_write_get_304_no_content_type(client, app):
         resp.status = HTTPStatus.NOT_MODIFIED
 
     await client.get('/test')
-    assert client.protocol.writer.data == b'HTTP/1.1 304 Not Modified\r\n\r\n'
+    assert client.protocol.transport.data == b'HTTP/1.1 304 Not Modified\r\n\r\n'
 
 
 async def test_write_get_1XX_no_content_type(client, app):
@@ -67,7 +67,7 @@ async def test_write_get_1XX_no_content_type(client, app):
         resp.status = HTTPStatus.CONTINUE
 
     await client.get('/test')
-    assert client.protocol.writer.data == b'HTTP/1.1 100 Continue\r\n\r\n'
+    assert client.protocol.transport.data == b'HTTP/1.1 100 Continue\r\n\r\n'
 
 
 async def test_write_head_no_content_type(client, app):
@@ -77,17 +77,7 @@ async def test_write_head_no_content_type(client, app):
         resp.status = HTTPStatus.OK
 
     await client.head('/test')
-    assert client.protocol.writer.data == b'HTTP/1.1 200 OK\r\n\r\n'
-
-
-async def test_write_connect_no_content_type(client, app):
-
-    @app.route('/test', methods=['CONNECT'])
-    async def connect(req, resp):
-        resp.status = HTTPStatus.OK
-
-    await client.connect('/test')
-    assert client.protocol.writer.data == b'HTTP/1.1 200 OK\r\n\r\n'
+    assert client.protocol.transport.data == b'HTTP/1.1 200 OK\r\n\r\n'
 
 
 async def test_write_cookies(client, app):
@@ -99,7 +89,7 @@ async def test_write_cookies(client, app):
         resp.cookies.set(name='name', value='value')
 
     await client.get('/test')
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert b'\r\nSet-Cookie: name=value; Path=/\r\n' in data
 
 
@@ -113,7 +103,7 @@ async def test_write_multiple_cookies(client, app):
         resp.cookies.set('other', 'value2')
 
     await client.get('/test')
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert b'\r\nSet-Cookie: name=value; Path=/\r\n' in data
     assert b'\r\nSet-Cookie: other=value2; Path=/\r\n' in data
 
@@ -127,7 +117,7 @@ async def test_write_cookies_with_path(client, app):
         resp.cookies.set('name', 'value', path='/foo')
 
     await client.get('/test')
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert b'\r\nSet-Cookie: name=value; Path=/foo\r\n' in data
 
 
@@ -141,7 +131,7 @@ async def test_write_cookies_with_expires(client, app):
                          expires=datetime(2027, 9, 21, 11, 22))
 
     await client.get('/test')
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert (b'\r\nSet-Cookie: name=value; '
             b'Expires=Tue, 21 Sep 2027 11:22:00 GMT; Path=/\r\n') in data
 
@@ -155,7 +145,7 @@ async def test_write_cookies_with_max_age(client, app):
         resp.cookies.set('name', 'value', max_age=600)
 
     await client.get('/test')
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert (b'\r\nSet-Cookie: name=value; Max-Age=600; Path=/\r\n') in data
 
 
@@ -168,7 +158,7 @@ async def test_write_cookies_with_domain(client, app):
         resp.cookies.set('name', 'value', domain='www.example.com')
 
     await client.get('/test')
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert (b'\r\nSet-Cookie: name=value; Domain=www.example.com; '
             b'Path=/\r\n') in data
 
@@ -182,7 +172,7 @@ async def test_write_cookies_with_http_only(client, app):
         resp.cookies.set('name', 'value', httponly=True)
 
     await client.get('/test')
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert (b'\r\nSet-Cookie: name=value; Path=/; HttpOnly\r\n') in data
 
 
@@ -195,7 +185,7 @@ async def test_write_cookies_with_secure(client, app):
         resp.cookies.set('name', 'value', secure=True)
 
     await client.get('/test')
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert (b'\r\nSet-Cookie: name=value; Path=/; Secure\r\n') in data
 
 
@@ -208,7 +198,7 @@ async def test_write_cookies_with_multiple_attributes(client, app):
         resp.cookies.set('name', 'value', secure=True, max_age=300)
 
     await client.get('/test')
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert (b'\r\nSet-Cookie: name=value; Max-Age=300; Path=/; '
             b'Secure\r\n') in data
 
@@ -224,5 +214,5 @@ async def test_delete_cookies(client, app):
 
     resp = await client.get('/test')
     assert resp.status == HTTPStatus.OK
-    data = client.protocol.writer.data
+    data = client.protocol.transport.data
     assert b'\r\nSet-Cookie: name=value\r\n' not in data