+++ docs/changelog.md
@@ -17,6 +17,8 @@ A changelog:
       a 404 is still automatically raised.
       ([#45](https://github.com/pyrates/roll/pull/45))
 * Added `request.host` shortcut ([#43](https://github.com/pyrates/roll/pull/43))
+* Introduced protocol upgrade and protocol configuration for routes. Two protocols are shipped by default : HTTP and Websocket ([#54](https://github.com/pyrates/roll/pull/54)).
+* The route is now resolved as soon as the URL has been parsed. In addition, the route lookup method has been split up from the application `__call__`method, to allow easy override.
 
 
 ## 0.8.0 - 2017-12-11